LPS-196642 Inline CommerceCartResourceUtil methods in CommerceCartResource
